Renamed service_dependency helper so it is not automatically included. Updated existing specs as needed.
